# uno-app/openapi.yaml
openapi: 3.0.3
info:
  title: UNO App API
  description: |
    API for UNO License Distribution Platform.

    ## Authentication
    Public endpoints require no authentication. Admin endpoints require HMAC-signed requests.
  version: 0.1.0
  contact:
    name: UNO Support
    url: https://unetwork.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.unetwork.io/api/v1
    description: Production
...
paths:
...
  /licenses/claim:
    post:
      tags: [Licenses]
      summary: Claim a license
      description: Claim a license using a lease code
      operationId: claimLicense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
      responses:
        '200':
          description: Claim result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # Statistics
...
components:
...
  responses:
    ServiceUnavailable:
      description: Database not available
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Database not available
            code: SERVICE_UNAVAILABLE
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - SERVICE_UNAVAILABLE
            - UNAUTHORIZED
            - NOT_FOUND
            - BAD_REQUEST
            - INTERNAL_ERROR
            - IMPORT_ERROR
...
    ClaimRequest:
      type: object
      required: [lease_code]
      properties:
        lease_code:
          type: string
          description: License lease code (format XXXX-XXXX-XXXX)
          example: ABCD-EFGH-IJKL
        device_id:
          type: string
          nullable: true
          description: Optional device identifier for binding
    ClaimResponse:
      type: object
      properties:
        success:
          type: boolean
        license:
          $ref: '#/components/schemas/LicenseDto'
          nullable: true
        license_id:
          type: string
          nullable: true
        license_key:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    LicenseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lease_code:
          type: string
        split_type:
          $ref: '#/components/schemas/SplitType'
        user_share:
          type: integer
        operator_share:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        claimed:
          type: boolean
    SplitType:
      type: string
      enum: ['50:50', '55:45', '60:40']
...